changing to read report mode instead of boot mode

+++ main/esp_hidd_prf_api.h
@@ -160,7 +160,7 @@ void esp_hidd_send_consumer_value(uint16_t conn_id, uint8_t key_cmd, bool key_pr
 
 void esp_hidd_send_keyboard_value(uint16_t conn_id, key_mask_t special_key_mask, uint8_t *keyboard_cmd, uint8_t num_key);
 
-void esp_hidd_send_mouse_value(uint16_t conn_id, uint8_t mouse_button, int8_t mickeys_x, int8_t mickeys_y, int8_t wheel);
+void esp_hidd_send_mouse_value(uint16_t conn_id, uint8_t mouse_button, int16_t mickeys_x, int16_t mickeys_y, int8_t wheel);
 
 #ifdef __cplusplus
 }
